fixed data transfer with scoring pages for coral
