+++ website/src/components/ScoringComponents/ScoringAlgae/ScoringAlgaeSection.jsx
@@ -4,51 +4,6 @@ import { useLocation } from "react-router-dom";
 import ScoringPickup from "../ScoringPickup";
 
 const ScoringAlgaeSection = ({ pickPositions, pickCounts, placeCounts }) => {
-  const location = useLocation();
-  const states = location.state;
-
-  // const pickPositions = ["Reef", "Mark 1", "Mark 2", "Mark 3"];
-
-  // const [pickReefCount, setPickReefCount] = useState(
-  //   states?.inputs?.pickStationCount || 0
-  // );
-  // const [pickMark1Count, setPickMark1Count] = useState(
-  //   states?.inputs?.pickMark1Count || 0
-  // );
-  // const [pickMark2Count, setPickMark2Count] = useState(
-  //   states?.inputs?.pickMark2Count || 0
-  // );
-  // const [pickMark3Count, setPickMark3Count] = useState(
-  //   states?.inputs?.pickMark3Count || 0
-  // );
-
-  // const pickCounts = [
-  //   { position: "Reef", count: pickReefCount, setCount: setPickReefCount },
-  //   { position: "Mark 1", count: pickMark1Count, setCount: setPickMark1Count },
-  //   { position: "Mark 2", count: pickMark2Count, setCount: setPickMark2Count },
-  //   { position: "Mark 3", count: pickMark3Count, setCount: setPickMark3Count },
-  // ];
-
-  // const [placeNetShot, setPlaceNetShot] = useState(
-  //   states?.inputs?.placeNetShot || 0
-  // );
-  // const [placeProcessor, setPlaceProcessor] = useState(
-  //   states?.inputs?.placeProcessor || 0
-  // );
-  // const [placeDropMiss, setPlaceDropMiss] = useState(
-  //   states?.inputs?.placeDropMiss || 0
-  // );
-
-  // const placeCounts = [
-  //   { position: "Net Shot", count: placeNetShot, setCount: setPlaceNetShot },
-  //   {
-  //     position: "Processor",
-  //     count: placeProcessor,
-  //     setCount: setPlaceProcessor,
-  //   },
-  //   { position: "Drop/Miss", count: placeDropMiss, setCount: setPlaceDropMiss },
-  // ];
-
   const [pickPositionSelected, setPickPositionSelected] = useState("");
 
   return (
