+++ website/src/components/ScoringComponents/ScoringCoral/ScoringCoralPlaceCounter.jsx
@@ -20,7 +20,7 @@ const ScoringCoralPlaceCounter = ({
           setPlaceCount(placeCount + 1);
           pickCounts.find((pickCount) => {
             if (pickCount.position == pickPositionSelected) {
-              pickCount.setCount(pickCount.count - 1);
+              pickCount.setCount(pickCount.count + 1);
             }
           });
           setPickPositionSelected("");
