+++ website/src/components/ScoringComponents/ScoringCoral/ScoringCoralSection.jsx
@@ -5,63 +5,6 @@ import ScoringCoralPlaceMap from "./ScoringCoralPlaceMap";
 import ScoringPickup from "../ScoringPickup";
 
 const ScoringCoralSection = ({ pickPositions, pickCounts, placeCounts }) => {
-  const location = useLocation();
-  const states = location.state;
-
-  // const pickPositions = ["Station", "Mark 1", "Mark 2", "Mark 3"];
-
-  // const [pickStationCount, setPickStationCount] = useState(
-  //   states?.inputs?.pickStationCount || 0
-  // );
-  // const [pickMark1Count, setPickMark1Count] = useState(
-  //   states?.inputs?.pickMark1Count || 0
-  // );
-  // const [pickMark2Count, setPickMark2Count] = useState(
-  //   states?.inputs?.pickMark2Count || 0
-  // );
-  // const [pickMark3Count, setPickMark3Count] = useState(
-  //   states?.inputs?.pickMark3Count || 0
-  // );
-
-  // const pickCounts = [
-  //   {
-  //     position: "Station",
-  //     count: pickStationCount,
-  //     setCount: setPickStationCount,
-  //   },
-  //   { position: "Mark 1", count: pickMark1Count, setCount: setPickMark1Count },
-  //   { position: "Mark 2", count: pickMark2Count, setCount: setPickMark2Count },
-  //   { position: "Mark 3", count: pickMark3Count, setCount: setPickMark3Count },
-  // ];
-
-  // const [placeL1Count, setPlaceL1Count] = useState(
-  //   states?.inputs?.placeL1Count || 0
-  // );
-  // const [placeL2Count, setPlaceL2Count] = useState(
-  //   states?.inputs?.placeL2Count || 0
-  // );
-  // const [placeL3Count, setPlaceL3Count] = useState(
-  //   states?.inputs?.placeL3Count || 0
-  // );
-  // const [placeL4Count, setPlaceL4Count] = useState(
-  //   states?.inputs?.placeL4Count || 0
-  // );
-  // const [placeDropMissCount, setPlaceDropMissCount] = useState(
-  //   states?.inputs?.placeDropMissCount || 0
-  // );
-
-  // const placeCounts = [
-  //   { position: "L1", count: placeL1Count, setCount: setPlaceL1Count },
-  //   { position: "L2", count: placeL2Count, setCount: setPlaceL2Count },
-  //   { position: "L3", count: placeL3Count, setCount: setPlaceL3Count },
-  //   { position: "L4", count: placeL4Count, setCount: setPlaceL4Count },
-  //   {
-  //     position: "Drop/Miss",
-  //     count: placeDropMissCount,
-  //     setCount: setPlaceDropMissCount,
-  //   },
-  // ];
-
   const [pickPositionSelected, setPickPositionSelected] = useState("");
 
   return (
